Tuggummi's Texer Resources
This is the complete set of Tuggummi's resources for texer. Not sure
whether you want to keep them in here or put them into a seperate file.
There are also other Texer packs available, although this is clearly
the most popularly used one.

I replaced your included images, as mine seemed to be slightly smaller.